changed FracX, FracY to floats

+++ Archipelago/Archipelago/NoiseGeneration.cpp
@@ -58,8 +58,8 @@ float NoiseGeneration::getInterpolatedNoise(float x, float z)
 	int intX = (int)x; //interger part of value x
 	int intZ = (int)z; //interger part of value z
 
-	int fracX = x - intX; //fraction/decimal section of the x value
-	int fracZ = z - intZ; //fraction/decimal section of the z value
+	float fracX = x - intX; //fraction/decimal section of the x value
+	float fracZ = z - intZ; //fraction/decimal section of the z value
 
 	 //calutating the height of the nearest points to the noise function
 	float v1 = getSmoothNoise(intX, intZ);
